add makefile target for logging into development database
currently the readme contains information on how to a two-step
docker command for authenticating with the running development
database container.

this commit shortens the step to a single command and moves the
command into a make target, meaning developers will only need to
execute 'make db/login' to get an interactive psql command prompt.

verification:
- if no database is setup, run 'make db/setup'
- run 'make db/login'
- ensure you now have an interactive prompt in the development
  database
- ensure the updated readme steps result in the same interactive
  prompt