add cluster creation command to avoid rest api requirement
currently, the cluster service is not accessible through the rest
api that the cli 'serve' command provides. this makes manual
testing of changes to the cluster creation logic difficult.

this commit exposes an additional command on the cli for cluster
service interaction. this allows the cluster service to be tested
manually without exposing a rest api route. this command leverages
the environment of the rest api, meaning database connections etc
are still used.

verification:
- ensure ocm and aws tokens have been exposed per the readme
- run 'make binary'
- run './managed-services-api cluster create'
- ensure a cluster is created and the command completes with an
  ok exit code
- ensure the readme additions are accurate