remove custom undefinedvariableerror error from service
currently, there is a custom error implementation for when
variables are undefined. however we can leverage serviceerror
type with the validation code for our current needs.

this change removes custom_errors.go, which contains this custom
type, and replaces all uses of the type with a validation service
error.

verification:
- run 'go test ./...'
- ensure tests complete successfully