update multiaz field in the kafka request to be a boolean
currently the multiaz field in the kafka request type is a string.
however as multi az is currently set to either true or false, we
can treat it as a boolean.

this commit performs a database migration from a text column to a
boolean column in the database during the cli 'migrate' command
and also updates internal types to use a boolean.

verification:
- from master, run 'make db/teardown'
- from master, run 'make db/setup'
- from master, run 'make binary'
- from master, run './managed-services-api migrate'
- ensure the database currently has 'text' as the column type in
  the kafka_requests table
- from this commit, run './managed-services-api migrate'
- ensure the database has 'boolean'/'bool' as the column type in
  the kafka_requests table
- run 'make test/run'