refactor openapi kafka details to include async flag
currently the post endpoint for creating kafka resources is async,
however we do not follow the ocm api guidelines of insisting on an
async boolean flag being provided and set to true to be considered
a valid request.

this commit adds a new async flag to the openapi spec of the
service to include the async flag.

this commit also removes the cluster_id attribute from the
returned kafka data.

verification:
- run 'make generate'
- run 'git diff' and ensure no files have changed
- run 'make run/docs'
- go to 'localhost'
- ensure the post kafkas endpoint has an async flag in the openapi
  spec
- ensure the kafka resource in the openapi spec does not contain a
  cluster id