create required credential files when building the test image
The aws and ocm credentials needs to be set up before running
the integration tests. The client ID and secret will need to have
dummy credentials. In order to facilitate this, a new make target
was added to set up the ocm client id and secret files with some
dummy value.

The make target for setting the ocm service token has now been
renamed to ocm/token/setup.