update the service templates
The service templates needed to be updated so that we can deploy our
service to an OpenShift cluster. The current templates are based on
the example ocm service and still contains references to that service.

The service templates were updated with missing information and name changes
from ocm example service to managed-services-api.

JIRA: https://issues.redhat.com/browse/MGDSTRM-369