add cloud provider configmap to the service template
The service requires a configmap that specifies the list of supported
cloud providers. Added the configmap to the service template to create it
when deploying the service.

This change came from a suggestion from App SRE so that it will be easier for
us to merge changes to this configmap for stage/production.

Related MR: https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/11739
JIRA: https://issues.redhat.com/browse/MGDSTRM-567