allow metrics host to be overridden in template
currently the metrics address we expose on cannot be overridden
from the template, this means in ocm it will default to
localhost:8080 which will cause metrics to not be exposed.

this commit ensures the metrics address can be overridden via the
existing template parameter.

this commit also fixes a small env naming issue, where AMS_ENV is
used instead of OCM_ENV