fix(kafka): delete kafkas in accepted status
A kafka in "accepted" status does not have a "ClusterID", thus attempting to delete the related
syncset would always fail thus resulting to a 404 status code.

Closes https://issues.redhat.com/browse/MGDSTRM-739

To verify this:

Before this patch, run the following the following to reproduce the
error:

1. Start with a clean database without any cluster i.e `make db/setup`
then  `./managed-services-api migrate`.
2. In case you had the database setup already, delete the existing
clusters. This is to make sure that the reconciler does not find
suitable clusters to provision kafkas. You can also, increase the
reconciler worker interval to a large number if you do not want to
delete clusters.
3. Start the API `./managed-services-api serve`
4. Create a kafka with `curl -v -XPOST -H "Authorization: Bearer $(ocm token)" http://localhost:8000/api/managed-services-api/v1/kafkas\?async\=true -d '{ "region": "us-east-1", "cloud_provider": "aws",  "name": "serviceapi", "multi_az": true}'`, note the created id.
5; Delete the created kafka, `curl -v -XDELETE -H "Authorization: Bearer
$(ocm token)"
http://localhost:8000/api/managed-services-api/v1/kafkas/<kafka-request-id>`
 and observe the delete failure.

To verify the fix:

Now apply this patch and run step (3) again.

Applying step (5) should successful delete the kafka.

You can verify that by checking that the kafka is deleted with `curl -v  -H "Authorization: Bearer $(ocm token)"
http://localhost:8000/api/managed-services-api/v1/kafkas/<kafka-request-id>`