fix: use db.update instead of db.save on Kafka create
db.Save updates the whole record in the database if found, otherwise it will create a new record.
This becomes an issue if the Kafka instance was deleted during creation. Since it cannot
find the record any longer, it will insert a new record to the table.

We should use db.Update instead. This will only set the values of fields that requires updating instead of the whole
record. This will avoid overwriting fields like status in case it gets outdated. If the record cannot be found, no
further action will be done.

JIRA: https://issues.redhat.com/browse/MGDSTRM-850