fix TestKafkaDelete_DeleteDuringCreation integration test race condition
Fixes a race condition in the TestKafkaDelete_DeleteDuringCreation
integration test where a race condition in test code made
LeaderElectionManager call Stop method before Start method
initialized the internal golang channel to control shutdown
of the LeaderElectionManager.

The additional goroutine used to start the LeaderElectionManager
in integration tests has been removed. There's no need for it
as the Start method of LeaderElectionManager itself creates its own
thread. Part of the Start method contents have been moved inside the
goroutine that it creates too to made them non-blocking for the
consumers of the LeaderElectionManager.

In this way, we are able to ensure that when running the integration
tests that make use of the LeaderElectionManager the channel is
initialised before calling Stop from the tests.