fix(allow-list): MGDSTRM-1052 individual allowed accounts unable to create Kafkas
Individual accounts listed as service accounts and part of the same org
were unable to create additional kafka once the total sum of kafkas for
that organisation was reached.

The fix was to not consider the organisationId under this case (for
accounts defined as service accounts) and always filter by owner.

Fixes https://issues.redhat.com/browse/MGDSTRM-1052