fix(kafka): MGDSTRM-1204  validate kafka name uniqueness across organisation or service account.
This verifies that a name is unique across organisation or owner of a given service account during
kafka creation. Fixes https://issues.redhat.com/browse/MGDSTRM-1204

This is a breaking change as it adds validation by returning 409 status
code in case the name exists already. We should notify the UI/CLI team
before merging this.

The error code is `MGD-SERV-API-36` and the reason is `Kafka cluster name is already used`.

To verify the change, make sure that the newly added tests are passing.