dynamic_scaling: modify cluster scaling criteria:
The implementation now assumes that resizeInfo.nodeDelta provided
by the agent will be always 3 and that resizeInfo.delta capacity
attributes will not change in value. In case that assumption
changes then the implementation will need adjustment.

The number of nodes to scale-up and scale-down is taken from
resizeInfo.nodeDelta.

The scale-up thresholds are the corresponding Kafka capacities for
connetions and partitions of a single kafka instance of model T.
Scale-up will be triggered when the remaining capacity is less than
any of the scale-up thresholds. The scaling will not be performed
if the scaling action would make the number of nodes to exceed
the (restricted) ceiling value

The scale-down thresholds are the corresponding Kafka capacities
specified in resizeInfo.delta. This means that the capacities
correspond to 3 OSD compute nodes.
Scale-up will be triggered when the remaining capacitiy is higher
than all of those thresholds. This means that a scale-down will
be triggered when remaining capacity is higher than 3 full OSD
cluster compute nodes. The scaling will not be performed
if the scaling action would make the number of nodes to be
below currentWorkLoadMin or the (restricted) floor value or
the MultiAZ limitations (3)