feat: implements ADR-00035 by storing connector secrets in a Vault service.
It identifies connector spec fields as containing a secret if it of an expected type.  The values of secret fields are replaced with vault references.  Our DB stores the references. When user get the connectors, those references are stripped out.

When the connectors are listed by the agent apis, the vault references are resolved and the secrets are encoded as base64 so that they can be installed into the target OSD clusters.