feat: remove cloud provider and region output in logs
The logic within this cloud providers loop was intended to be changed to
add logic for automating the creation of new OSD clusters on every supported
cloud provider and region. This has been added into the reconcileClustersForRegions instead,
leaving this loop to be unused. This has been removed to reduce calls to ocm as well
as reduce log noise.