fix(integration): osd cluster setup should call the agent-cluster endpoint to update it to ready
KAS fleetshard operator should now be enabled for all environments so the current
way of waiting for the cluster to be ready will no longer work. A new function has
been added to wait for and update the osd cluster status to a 'ready' state via
the agent-cluster endpoint. This should be used for setting up an osd cluster for
the cluster manager tests.