test(integration): change cluster status criteria when polling for clusters
Clusters will never be in a 'ready' state until the agent-cluster endpoint is called.
Instead of waiting for 'ready' clusters, it should just check for clusters in a 'waiting_for_kas_fleetshard_operator'
status (ensuring that a cluster is created via manual scaling does not need it to be in a 'ready' state, ensuring that it is
registered in the database and is processed by the cluster worker is enough).

Existing clusters created for this test are updated to 'ready' later on in the test to test the kafka placement
strategy.