fix(integration): ensure kafka status are updated to ready and deleting
Since the KAS fleetshard operator cannot call the agent clusters endpoints when running
integration tests, we need to update the status of Kafka requests ourselves if they need
to be in a 'ready' or 'deleting' state.

These changes ensures that the agent cluster kafka status endpoint is called. Kafkas with
deprovisioning status will be updated to 'deleting'. Kafka requests in a 'provisioning',
'failed' and 'ready' state will be updated to 'ready'.