fix: don't skip the http stack, the RequestURI was not being set which would break the error handing.
We likely need to avoid this anyways since 1 graphql request can generate many openapi http requests, and in staging/prod we will want those to hit the rate limiting proxy to avoid abuse.
THe graphql endpoint is now enabled by setting the --graphql-api-url, which should be the rate limiting proxy address in a normal deployment.