chore: add comments where code needs to be removed/refactored to remove mas-sso auth support for /federate
MAS SSO auth support for the /federate endpoint is only a temporary solution and needs to be removed once we
migrate to sso.redhat.com. Comments are added to note what needs to be removed and updated to do this. This
will be done as part of https://issues.redhat.com/browse/MGDSTRM-6159