feat: only return regions that support the specified instance_type
The query parameter 'instance_type' was added for the cloud provider
regions endpoint to filter the regions by supported Kafka instance type.

The previous behaviour was not ideal as it changes the meaning of
'enabled' depending on the value of the query parameter. Instead of
this, we should actually filter the region list instead:
  - If specified, only regions that support the specified Kafka instance
    type should be returned.
  - If not specified, all regions should be returned.

Related JIRA: https://issues.redhat.com/browse/MGDSTRM-6492