fix(admin): only compare against actual versions when they are present
Previously, when looking to update the Kafka that is not yet in a ready state (Actual versions are
only reported when a Kafka is being provisioned), the update failed because of comparision issue.
Since we allow to update kafka in preparing, provision state (see
https://github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/blob/main/internal/kafka/constants/kafka.go#L91-L98
and
https://github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/blob/main/internal/kafka/constants/kafka.go#L91-L98)
it make sense to only check against actual versions when they are already set in the database. The
setting of the actual kafka versions happens at
https://github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/blob/354878177fcd13b5ca6cd9dae242bc567eb07a8e/internal/kafka/internal/services/data_plane_kafka.go#L137