feat: update kafka config module to use the yaml.v2 pkg
Updated the Kafka Config Module to use gopkg.in/yaml.v2 instead of
ghodss/yaml as this is mainly used, making it consistent with the rest
of the project.

yaml.v2 also has the function UnmarshalStrict which is useful for
catching additional errors like duplicate mapping or unsupported keys
in the data.