chore(deps): update gocloak to v11 to fix CVE-2020-26160
This fix the Authorization bypass in github.com/dgrijalva/jwt-go issue.

jwt-go allows attackers to bypass intended access restrictions in
situations with []string{} for m["aud"] (which is allowed by the
specification). Because the type assertion fails, "" is the value of
aud. This is a security problem if the JWT token is presented to a
service that lacks its own audience check. There is no patch available
and users of jwt-go are advised to migrat to golang-jwt at version 3.2.1
which we already have in our go.mod