chore(ams): provide default value for ams url
AMS is used for account management. When the url is empty, it creates a 'nil' connection. When
trying to use the services that relies on ams being properly connected e.g account_services and
ams-quota services, this causes a panic error.

This is only one part of the fix.
The proper fix will be to:
- have a default implementation of the account service when ams is not
in use
- properly validate the ams-url and all the other flags when ams service
is being used