refactor: remove eval to prod kafka instance type migration
Migrating eval to developer would result in instances that have the
developer type but eval brokers and limits. This makes capacity calculations hard.
We remove the migration and we will let existing eval instances expire.
Code that deals with both eval and dev instances will need to coexist for some
time.