fix: add additional labels to differentiate metric results for federate endpoint
The following changes fixes the issue when collecting and converting
user facing metrics to Prometheus format for the /federate endpoint.
This was caused by the same metric returning multiple results with the
same values for its labels.

Additional labels were added to differentiate each metric result. These
labels were also added to the results of /query and /query_range
endpoint to keep consistency.