fix: allow multiAZ criteria to be configurable on cloud providers capacity reporting
Only 'Standard' Kafka instances should have the criteria of 'multi_az:
true'. 'Developer' Kafka instances can be scheduled either in a single
az or multi az cluster.

Gorm ignores a boolean criteria that is not set or has a false value.
Because of this, MultiAZ does not need to be set in the criteria when
checking capacity for 'Developer' Kafka instance type.