fix(clusters): do not wait for operators to become ready to store FSO sync client_id/client_secret
1. As reported in https://issues.redhat.com/browse/MGDSTRM-8409, for a new
OSD cluster, the fleet-manager will report 404 until the cluster is full
terraformed. These errors make it hard to investigate issues as they
could be misleading. So let's make sure that the client_id and
client_secret are stored in the database as soon as they are available
to avoid these 404 errors from being sent to the fleetshard sync.

2. By storing the client_id, client_secret as soon as they are
available and not wait for the operators installation readiness which
can take a few minutes, we are sure to only generate the client_id and
client_secret only once. This avoids cases where we attempted to
generate the client each time until the operators where ready because
the client weren't stored in the database. This creates an uncessary
load on the sso provider and for some providers that do not support the
ability to see if the client exists already, it could lead to create
some phantom accounts