test: split and fix dataplane endpoint get and update managedkafkas integration test
The GetAndUpdateManagedKafkas integration test for the data plane
endpoint creates Kafkas in a 'preparing' state. However, depending on
the timing of the reconciliation, the 'preparing' Kafkas may be updated
to the next state. This causes an issue for the assertion that checks
the number of Kafkas returned from the GetManagedKafkasByClusterId endpoint
as it is expected to not include the 'preparing' Kafkas in its response.

The test has been split so that the UpdateManagedKafka tests are not
affected by the increase in the reconcile repeat interval to prevent
'preparing' Kafkas from moving onto the next state.

The following changes also includes a refactor of how the test Kafkas
are defined. It should now use the builder functions to keep it consistent
with the rest of the integration tests.