fix: remove check for EnableAuthenticationOnKafka from kafka deletion reconcile
This check causes all Kafkas that are marked as deprovision to be
removed immediately without waiting for fleetshard to report that
resources on the data plane cluster has been deleted for clusters. This
check has been removed to prevent this.

Kafkas will only move from 'preparing' to 'provisioning' if the boostrap
server host is set. Therefore the only check we should do is to see if
the bootstrap server host is set. This would mean that resources haven't
been created on the data plane cluster yet so we can remove the kafka
immediately.