feat: make data plane image pull secret file optional
The data plane image pull secret file should only be available when the
images used by the fleetshard and strimzi operators lives in a private
repository. If these images are public, the image pull secret does not
need to be available.

We should not fail kas fleet manager start up when the image pull secret
file does not exist. A warning log message should be shown instead.