feat: add a dedicated model to represent connector namespace within private api
This represent a first step toward having a fully dedicated private API
to deal with connectors namespaces and reduce the risk to introduce
breaking changes for the agen as conseguences fo public APIs changes.

This commit is meant to be backward compatible and introduces a set of
of additional component schemes to be used only in the private agent
APIs.

[1] https://issues.redhat.com/browse/MGDCTRS-877