feat: add max_data_retention_size to the public api KafkaRequest schema
Adds max_data_retention_size to the KafkaRequest schema of the KAS Fleet
Manager public openapi spec.

The max_data_retention_size field shall contain the value of the maximum data
storage available to a Kafka instance. This should be in a bytes format,
similar to how its returned from the /instance_types endpoint.

The value used for this is retrieved from the field kafka_storage_size
from the kafka_requests table which stores the value as a string Quantity
format. The Kafka Request presenter should handle the conversion from this
type to bytes.