feat(dynamic-scaling): add cluster capacity reporting
Closes https://issues.redhat.com/browse/MGDSTRM-8741

The capacity info object has the following structure:
```yaml
instance_type:
  max_nodes: <integer>
  max_units: <integer>
  remaining_units: <integer>
```

for example:
```json
{
 standard: {
   max_nodes: 30
   max_units: 10
   remaining_units: 4
 },
 developer: {
   max_nodes: 1,
   max_units: 10,
   remaining_units: 8
 }
}

```

1. To retrieve the information:
```go
cluster.RetrieveDynamicCapacityInfo()
```

the above method has to be called every time you want to check the
information.

2. To set the info

```go
cluster.SetDynamicCapacityInfo(map[string]DynamicCapacityInfo{})
```

NOTE:
1. The reconciler makes sure that when the scaling mode is not auto, the
object is reset to empty.
2. When the scaling is auto, the object is set initialised again with
the content of `maxNodes` read from the config. The assumption is that,
the SOP will include the necessary steps on how the corresponding
machine pool will be created because the reconciler does not handle the
auto creation of machine pools for `ready` clusters
3. Node info are only sent to the data plane when scaling is auto.
Otherwise an empty object is sent.
4. On each ManagedKafkaAgentCR status update, we update the capacity
info. This is because `remainingUnits` is a dynamic info and changes all
the time. Also, this caters for situation where the reported `maxUnits`
has changed for one reason or the other e.g deployment topology changes
etc