fix(metrics): reset capacity used and capacity available metrics for KafkaManager reconciler
The two metrics were calculated in the KafkaManager reconciler but the
metrics were reset from the clusters reconciler.

By moving this to the correct resets methods, the metrics will get properly reset when the leader changes for the KafkaManager reconciler.
Which in turn will ensures that the old leader pod does not show stale data when it's scraped