feat: take cluster max streaming unit into account for first ready cluster strategy
The maxUnit in the dynamic_capacity_info of a cluster object represents
the maximum Kafka streaming unit that is available in the data plane
cluster.

Taking this into account in the first read cluster placement strategy
provides a reliable way of minimising the amount of rejected Kafka
Requests due to a cluster not having enough capacity left as this value
does not change often.

This also includes changes to move the
CountStreamingUnitByRegionAndInstanceType from the kafka service to the
cluster service to avoid a dependency cycle with the cluster placement
strategy and kafka service. This function has also been re-named to
FindStreamingUnitCountByClusterAndInstanceType to be consistent with the
functions in the cluster service.