feat: add minimum available capacity slack concept
As part of dynamic scaling, the concept of a minimum
available capacity slack in streaming units has been introduced.
The available capacity slack of a supported instance type in a region
is the desired minimum number of streaming units that should
be available(free) at any moment for that instance type in the region.

It has been introduced in the provider configuration file of the
fleet manager. There is one capacity slack value per defined instance
type in the region.
The attribute is optional. If it is not set then it is interpreted
as a capacity slack of 0 is desired.
The capacity slack is intended to be used only when dynamic scaling
is enabled.