test: use a dummy data plane cluster for kafka create integration tests
Instead of using the existing data plane cluster created by the cluster
manager integration test or creating a new one, the kafka create tests
should use a dummy data plane cluster added directly to the database.

This will save time when running these tests as we do not need to wait
for the cluster to be ready.