feat: change FindStreamingUnitCountByClusterAndInstanceType to retrieve all but failed clusters
Before this change the behavior of the FindStreamingUnitCountByClusterAndInstanceType method
was to initially retrieve all custers that don't have a ClusterID. This is, to retrieve all clusters
but those that are in 'accepted' state.
This change modifies that criteria and it now retrieves all clusters but those that are
in 'failed' state.
The reason of this change is due to we want to return the information for all clusters
status with the exception of the failed ones so the consumers of the method can decide to
filter the ones they want based on the status.