refactor(clusters_providers): add supported cluster providers types in the provider factory
This avoids the need to enlist supported clusters providers from the database all the time. The list
of provider types is known at compilation time, we can just return it wherever it is needed. The
advantage of this approach is that it saves us a few query to the database and the roundtrip.
Another advantage is that we do not need to have at least a cluter from the given cluster provider
to have a functional cloud providers endpoints