feat: expose and add GCP credentials configuration
Expose configurability of GCP credentials as a secret file.
For local development with a local process a file named
gcp.api-credentials is expected in the secrets directory.
For development in an OSD cluster it is created as the
gcp.api-credentials in the K8s secret named 'kas-fleet-manager'
from the secrets template. Its value can be configured through
the GCP_API_CREDENTIALS template parameter.
The default location of the credentials file can be modified
by using the newly added flag '--gcp-api-credentials-file' which
by default it is set to 'secrets/gcp.api-credentials'.