perf(metrics): distribute metrics queries
The metrics queries can be processed in parallel to further improve the metrics latency time. At the
moment the processing is done sequentially meaning which means that the latency will be the total
time it takes for all sequential calls to observatorium to finish. Doing it in parallel means that
the time it now takes will be the largest time it takes for one operation to complete