feat: add support for suspended status reported by the data plane
A new 'suspended' state is sent by the data plane which is determined
when the ManagedKafka CR Ready condition reason is set to 'Suspended'.

Additional states have also been added to the KafkaRequest. These will
need to be accommodate for when processing the real Kafka deployments.
- 'resuming': when a Kafka instance is in this state, the Ready
  condition reason will be set to 'Installing' and status to 'False'
- 'suspending': set by the user. Status of ManagedKafka CR is not
  changed prior to transitioning to 'Suspended' so kas fleet manager
must never overwrite it with any other status apart from 'suspended',
'resuming' or 'deprovision'.