feat: set route53 regional endpoints per cloud provider
When kafka instances are created from cloud providers different than
AWS AWS Route 53 is used to create DNS entries when external names
are enabled.
This changes the logic of the methods that make use of Route53 to
make sure a valid AWS region is specified, no matter the cloud
provider of the kafka instance being managed.
We define a AWS Route53 region for each cloud provider, independently
on the cloud provider's region where the kafka is allocated.
This is acceptable due to most Route53 features are global and
not region-specific. The region in those case is used merely
to access to a regional AWS API endpoint. We are currently not using
regional features of Route53 so we can use the implemented approach.