feat: set restrictive security contexts in KFM deployments
Starting from K8s 1.25 a new K8s built-in admission controller
called Pod Security Admission Control will allow cluster admins
to enforce K8s Pod Security Standards with Namespace Labels.
Additionally, OpenShift Security Context Constraints (SCC)
are in the progress of changing to align with the new criteria
to admit workloads according to the K8s Pod Security Standards
definitions.
With the introduction of the new built-in admission controller
that enforces K8s Pod Secuity Standards three policies belonging to
those standards are defined. One of those policies is the Restricted
policy.

As part of the changes done in OpenShift in the SCC to address the
new needs new SCC policies have been defined starting in OpenShift 4.11.
Specifically, the changes made align the deployments to comply with
the new introduced SCC policy named restricted-v2 which is the most
restrictive and it is the policy assigned by default to the
namespaces created by non-privileged users.

The set of changes change the KFM deployments to align with the
new SCC restricted-v2 policy. This:
- Removes some of the warnings that started to appear when deploying
  KFM starting in OCP 4.11 in relation to the security context not being
  configured to comply with the restricted-v2 SCC policy
- Ensure that KFM can be run with the restricted-v2 SCC policy.
  This allows non-privileged users to deploy KFM in OCP 4.11 and newer
  versions where admins decide to enforce restrictions according to the
  K8s Pod Security Standards
- Ensure that KFM can run in K8s clusters where cluster admins
  decide to enforce restrictions according to the K8s Pod Security
  Standards (with some caveats in some K8s versions as explained below)
- Follows best practices around giving the least amount of privileges
  needed for the workloads and follows K8s Pod Standard Security
  practices
- Make explicit the security needs of the KFM workloads

The set of changes in this commit KFM require OCP 4.11 as the minimum OCP version
used by KFM as the control plane. This is due to that in OCP 4.10
and earlier versions the SCC restricted policy does not allow to define
the securityContext seccompProfile attribute required by the K8s restricted
policy.