feat: set data plane openshift version to 4.11.22
The Fleetshard Operator addon is currently incompatible with OpenShift 4.12.
See https://issues.redhat.com/browse/MGDSTRM-10450 for further information.

This is a temporary change to ensure that our nightly tests do not fail on Fleetshard Operator
addon installation.

Once the issue is resolved, we can reset this back to an empty string so that KAS Fleet Manager
can provision data plane clusters with the latest version.

OpenShift 4.11.22 is the latest version of 4.11 currently available on OCM.