feat(certs-management): automatic certificate management
1. Add the capabability to automaticaly generate a wildcard certificate
   per standard kafka instance. Existing Kafkas will continue to share
the wildcard certificate but it'll be automatically managed. Developer
Kafkas will share one wildcard certificate under the
`*.trial.<kafka-domain-name>`.

2. Implement memory and Vault storage based on AWS Secret Manager. The
   in memory is useful for testing. The vault storage or file storage
based are useful when we need persistent storage. The vault one should
be used when we want to securely store the secrets.

Closes https://issues.redhat.com/browse/MGDSTRM-10444 and https://issues.redhat.com/browse/MGDSTRM-10443

This is highly based on the ADR and AP opened in this PR - https://github.com/bf2fc6cc711aee1a0c2a/architecture/pull/84