feat: reorganize README documentation and expand documentation
Simplify the main README documentation. The idea is that the main
README provides the most useful information first and links to
other relevant documents where more detail is provided in case
the user is interested.

Some additional documents have been created, being of special
relevance the populating configuration document and the
interacting with fleet manager document.