feat(certs): indicate that a Kafka can be paused for reconciliation in the data plane via the ManagedKafka CR annotation.
Previously when we fail to fetch the certificate content of a given Kafka we'll return an error and fail the process of generating ManagedKafka of a given data plane.
This means that we'd stop the reconciliation of all the Kafkas in that data plane as the Fleetshard sync was receiving an error each time.
This PR with the help of
https://github.com/bf2fc6cc711aee1a0c2a/kas-fleetshard/pull/883 (an
enhancement in the data plane) now makes it possible to indicate that a
given ManagedKafka that has encountered an error can be paused paused
from reconciliation. What this effectively means to the data plane is
that:
- The local ManagedKafka in the data plane will continue to be seen as
  the source of truth
- The desired ManagedKafka that has been paused for reconciliation will
  be skipped and won't update/overwrite the local ManagedKafka. Unless otherwise
  it is marked for deletion, in which case the Kafka is going to be deleted.

In order to do this, we needed a way to indicate to the data plane that
the MK can be paused for reconciliation: "managedkafka.bf2.org/pause-reconciliation".

We do this when an error occurs, which for now is when we fail to get the certificate of the Kafka.
The data plane then won't reconcile the local state of the ManagedKafka.
This ensures that at any time, we are able to reconcile all the Kafkas in a given cluster and
not error when we fail to build the Managedkafka CR of a given Kafka - which can potential impact
the reconciliation of others Kafkas.

Closes https://issues.redhat.com/browse/MGDSTRM-10919

This has to wait for https://github.com/bf2fc6cc711aee1a0c2a/kas-fleetshard/pull/883 to be
released.