chore: update github CODEOWNERS to use cos-fleet-manager-reviewers group
Use a GitHub group to notify connectors related code changes to
the team members of the connectors team instead of having to
specify individual people.
This allows us not having to add, ping and remember specific
people from the connectors team, remembering their
usernames and that group can also be updated dynamically
when people/join the connectors team.